com.unity.terrain-tools@3.0.2-preview.2 ## [3.0.2-preview.2] - 2020-09-28 ### Added - Editor Analytics integration. - Fallback shader to the HDRP Visualization shader. - PNG/TGA heightmaps support level remapping and vertical flipping. - Create New Terrain supports heightmaps in Texture2D format.
### Changed
- Changed the package name that's used to get core Universal Render Pipeline (URP) HLSL code from Lightweight to Universal.
- Added missing Terrain visualization settings.
- Removed Filter Stack Image assets.
- Changed Base Map Maximum Distance to 20000 to match the range in the Terrain Inspector.
- Terrain Toolbox splatmap modifications are now applied to Terrain only after you click **Apply To Terrain** in the Terrain Toolbox.

### Fixed
#### Brushes
- Brush sizes are clamped to a maximum based on texture size. ([1276290](https://issuetracker.unity3d.com/product/unity/issues/guid/1276290/))
- Clone tool correctly blends heightmap.
- Clone tool clones alphamaps as well as height.
- Clone tool correctly maintains clone position when you rotate the Scene Camera or click on the UI.
- Corrected Smooth tool behavior when you paint near the edges of Terrain tiles so that it produces less discontinuities. ([1186005](https://issuetracker.unity3d.com/issues/terrain-tools-terrain-gets-terraced-when-smoothed-near-the-edge-of-the-terrain-tile))
- Removed artifacts that occurred when you used the Twist Brush. ([1276448](https://issuetracker.unity3d.com/issues/terrain-tools-twist-tool-certain-values-cause-value-spikes-in-the-heightmap))
- Added support for Brush Mask Filters to the Wind Erosion tool.

#### Splatmaps
- Corrected blending when you apply splatmaps from the Terrain Toolbox.
- Fixed error thrown when you highlight a splatmap in the Terrain Toolbox.
- Fixed Undo/Redo functions after you apply a splatmap to a Terrain from the Terrain Toolbox.
- Fixed error thrown when you export a splatmap to a Terrain without selecting a Terrain. ([1240314](https://issuetracker.unity3d.com/issues/terrain-tools-terrain-toolbox-splatmap-export-to-terrain-throws-dividebyzeroexception-when-no-terrain-is-selected))
- Fixed Flip Adjustment on splatmap. ([1276204](https://issuetracker.unity3d.com/issues/terrain-tools-utilities-splatmap-flip-horizontal-and-flip-vertical-are-reversed))

#### Heightmaps
- Corrected heightmap when you create Terrain using a RAW file type in the Terrain Toolbox.
- Fixed Flip Axis when you use Import Heightmap to create Terrain.
- Heightmaps can be exported in 8-bit mode from the Terrain Toolbox. ([1276244](https://issuetracker.unity3d.com/issues/terrain-tools-toolbox-utilities-export-heightmaps-heightmap-depth-always-resets-to-16-bit))
- Terrain Toolbox heightmap export properly remaps levels.

#### Noise Editor
- Noise Editor displays correctly when you use the Personal (light) theme.
- Noise shader import works correctly on Linux. ([1188556](https://issuetracker.unity3d.com/issues/linux-adding-terrain-tools-package-makes-an-endless-loop-of-importing))
- You can now open only one Export Noise window at a time. ([1269540](https://issuetracker.unity3d.com/issues/terraintools-new-instance-of-export-noice-to-texture-window-is-created-every-time-even-though-another-is-already-open))
- Removed non-functional normalize option from the Export Noise UI.
- Unique asset path is generated when you export noise.
- Removed shader warnings from noise shaders.

#### Terrain Splitting
- Details and Trees are preserved when you split Terrain in the Terrain Toolbox. ([1248489](https://issuetracker.unity3d.com/product/unity/issues/guid/1248489/))
- Splitting Terrain with resolution of 33x33 warns the user and allows for an alternate split mode. ([1276273](https://issuetracker.unity3d.com/issues/terrain-tools-toolbox-utilities-splitting-terrain-with-heightmap-resolution-33-breaks-terrain-and-causes-error))
- Scenes now don't automatically save when you split Terrain in the Terrain Toolbox.

#### Additional Fixes
- Editing the Remap Curve in the Image Filter Stack no longer throws error. ([1254251](https://issuetracker.unity3d.com/issues/terraintools-nullreferenceexception-is-thrown-on-editing-the-remap-curve-in-the-image-filter-stack))
- Reverting to values in a Terrain preset now correctly reverts to saved values.
- Base texture size now appears correctly in the Terrain Inspector drop-down menu for Terrains you create with the Terrain Toolbox.
- Fixed toggle button behavior in the Terrain Toolbox, Brushes, and Brush Mask Filters.

+++ Editor/TerrainTools/BrushControllers/IBrushSmoothController.cs
@@ -6,7 +6,7 @@ namespace UnityEditor.Experimental.TerrainAPI
 	public interface IBrushSmoothController
 	{
 		bool active { get; }
-        float kernelSize { get; set; }
+        int kernelSize { get; set; }
 
 		void OnEnterToolMode();
 		void OnExitToolMode();
